move splat files to splat folder
