+++ src/splat/sort-worker.ts
