+++ src/splat/splat-resource.ts
